Remove icc warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@100928 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: b049924626a9ef834b2ff732616abccfcf63792c

+++ Online/Online/GaudiUPI/GaudiUPI/DialogDisplay.h
@@ -28,8 +28,8 @@ public:
   int repaint();
   int endUpdate();
   int beginUpdate();
-  int putString (const std::string& text,int video);
-  virtual int init (int num_col,int num_row, AbstractMainDisplay& parent);
+  int putString(const std::string& text,int video);
+  virtual int init(size_t num_col,size_t num_row, AbstractMainDisplay& parent);
 protected:
   DialogSubDisplay (AbstractMainDisplay& source);
   DialogSubMenu  *window() {     return m_window;   }
